helper-scripts/vis-seeds/requirements.txt
mako >= 1
pyyaml >= 5
wand >= 0.6.1
